refactor(ontology): moved modeling formats to onology-extra library
