+++ libs/datamodel/common/src/lib/common/index.ts
@@ -15,7 +15,6 @@ export * from './operating-system-type';
 export * from './package-repository';
 export * from './url';
 export * from './funding';
-export * from './modeling-formats';
 export * from './simulator-id-name-map';
 export * from './error';
 export * from './meta';
